Move PSF utils tests
